listing_builder/backend/requirements.txt
# /Users/shawn/Projects/ListingBuilderPro/listing_builder/backend/requirements.txt
# Purpose: Python dependencies for Marketplace Listing Automation Backend
# NOT for: Frontend dependencies or development tools

# FastAPI Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart>=0.0.7

# Database
supabase==2.3.4
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# AI/LLM
groq==0.4.1
google-generativeai>=0.8.0
openai>=1.50.0

# Validation & Serialization
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator>=2.0.0  # WHY: Required by Pydantic EmailStr

# HTTP Client
httpx==0.25.2
requests==2.31.0
curl_cffi>=0.7.0  # WHY: Chrome TLS impersonation — Amazon blocks httpx fingerprint

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2023.3

# Template Parsing (Amazon XLSM / eBay XLSX)
openpyxl==3.1.2

# Web Scraping (Allegro product pages)
playwright==1.41.0
playwright-stealth==1.0.6
# After pip install: run `playwright install chromium`

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Rate Limiting
slowapi==0.1.9

# Background Scheduling (lightweight, no external process)
apscheduler==3.10.4

# Payments
stripe==8.0.0

# Security
defusedxml>=0.7.1  # WHY: Prevents XXE attacks in RSS/Atom XML parsing

# Image Generation (A+ Content infographics)
Pillow>=10.0.0

# Logging
structlog==24.1.0
PyJWT[crypto]==2.8.0  # WHY: [crypto] needed for ES256 (Supabase JWKS verification)
